Moved function for loading plugins to a service, so that we can use dependency injection.
