+++ FrontEnd/Startup.cs
@@ -1,4 +1,5 @@
-using Api.Core.Helpers;
+using Api.Core.Interfaces;
+using Api.Core.Services;
 using FrontEnd.Core.Interfaces;
 using FrontEnd.Core.Models;
 using FrontEnd.Core.Services;
@@ -55,12 +56,9 @@ namespace FrontEnd
             Log.Logger = new LoggerConfiguration()
                 .ReadFrom.Configuration(Configuration)
                 .CreateLogger();
-
-            this.webHostEnvironment = webHostEnvironment;
         }
 
         public IConfiguration Configuration { get; }
-        private readonly IWebHostEnvironment webHostEnvironment;
 
         // This method gets called by the runtime. Use this method to add services to the container.
         // For more information on how to configure your application, visit https://go.microsoft.com/fwlink/?LinkID=398940
@@ -100,11 +98,10 @@ namespace FrontEnd
                 options.SerializerSettings.Converters.Add(new StringEnumConverter());
             });
 
-            // Load plugins for GCL and Wiser.
-            TypeHelpers.LoadPlugins(Configuration.GetValue<string>("FrontEnd:PluginsDirectory"), webHostEnvironment);
 
             // Setup dependency injection.
             services.AddHttpContextAccessor();
+            services.AddTransient<IPluginsService, PluginsService>();
             services.AddTransient<IBaseService, BaseService>();
             services.AddTransient<IImportsService, ImportsService>();
             services.AddTransient<IFrontEndDynamicContentService, FrontEndDynamicContentService>();
@@ -114,7 +111,7 @@ namespace FrontEnd
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IWebHostEnvironment env, IWebPackService webPackService)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env, IWebPackService webPackService, IPluginsService pluginService)
         {
             if (env.IsDevelopment())
             {
@@ -142,6 +139,9 @@ namespace FrontEnd
             });
 
             webPackService.InitializeAsync();
+
+            // Load plugins for GCL and Wiser.
+            pluginService.LoadPlugins(Configuration.GetValue<string>("Api:PluginsDirectory"));
         }
     }
 }
